Refactorized the Swizzling methods to be reused on different tests classes
